Added missing exception handler when accessing owner of a non-locked file lock. Deleting transactions now checks for the txn folder to exist

+++ aksw-commons-txn/src/main/java/org/aksw/commons/txn/impl/TxnMgrImpl.java
@@ -248,7 +248,9 @@ public class TxnMgrImpl
                 MoreFiles.deleteRecursively(lockRepoRootPath);
             }
         } finally {
-            MoreFiles.deleteRecursively(txnBasePath);
+            if (Files.exists(txnBasePath)) {
+                MoreFiles.deleteRecursively(txnBasePath);
+            }
         }
     }
 }
